+++ src/main/java/com/cooksys/twitter_api/repositories/UserRepository.java
@@ -2,12 +2,12 @@ package com.cooksys.twitter_api.repositories;
 
 import com.cooksys.twitter_api.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByCredentialsUsername(String username);
 
     boolean existsByCredentialsUsername(String username);
